Fixed some of the warnings when compiling with /W4

+++ Core/BF909x.h
@@ -8,8 +8,8 @@ private:
 	bool _bf9097Mode = false;  //Auto-detect for firehawk
 
 protected:
-	virtual uint32_t GetPRGPageSize() { return 0x4000; }
-	virtual uint32_t GetCHRPageSize() {	return 0x2000; }
+	virtual uint16_t GetPRGPageSize() { return 0x4000; }
+	virtual uint16_t GetCHRPageSize() {	return 0x2000; }
 
 	void InitMapper() 
 	{
