+++ Core/MMC3.h
@@ -140,8 +140,8 @@ class MMC3 : public BaseMapper
 		}
 
 
-		virtual uint32_t GetPRGPageSize() { return 0x2000; }
-		virtual uint32_t GetCHRPageSize() {	return 0x0400; }
+		virtual uint16_t GetPRGPageSize() { return 0x2000; }
+		virtual uint16_t GetCHRPageSize() {	return 0x0400; }
 
 		void InitMapper() 
 		{
