+++ Core/VRC2_4.h
@@ -30,8 +30,8 @@ class VRC2_4 : public BaseMapper
 		bool _hasIRQ;
 
 	protected:
-		virtual uint32_t GetPRGPageSize() { return 0x2000; }
-		virtual uint32_t GetCHRPageSize() {	return 0x0400; }
+		virtual uint16_t GetPRGPageSize() { return 0x2000; }
+		virtual uint16_t GetCHRPageSize() {	return 0x0400; }
 
 		void InitMapper() 
 		{
